fix: json number is double

+++ LineSharp/Messages/RichMenu.cs
@@ -63,15 +63,15 @@ namespace LineSharp.Messages
     public class Bounds
     {
         [JsonProperty("x")]
-        public int X { get; set; }
+        public double X { get; set; }
 
         [JsonProperty("y")]
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         [JsonProperty("width")]
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         [JsonProperty("height")]
-        public int Height { get; set; }
+        public double Height { get; set; }
     }
 }
